Remove casts

+++ rtsp/RTSP_options.c
@@ -43,7 +43,7 @@ static int send_options_reply(RTSP_buffer * rtsp)
         get_stat(200), cseq);
     strcat(r, "Public: OPTIONS,DESCRIBE,SETUP,PLAY,PAUSE,TEARDOWN,SET_PARAMETER" RTSP_EL);
     strcat(r, RTSP_EL);
-    bwrite(r, (unsigned short) strlen(r), rtsp);
+    bwrite(r, strlen(r), rtsp);
 
     fnc_log(FNC_LOG_CLIENT, "200 - - ");
